requirements.txt
fastapi
uvicorn
librosa
numpy
requests
boto3
python-dotenv